+++ CHANGELOG.md
@@ -45,7 +45,7 @@
 - First stable release
 
 
-[unreleased]: https://github.com/mayfield/syndicate/compare/v1.4.0...HEAD
+[unreleased]: https://github.com/mayfield/syndicate/compare/v2...HEAD
 [2]: https://github.com/mayfield/syndicate/compare/v1.4.0...v2
 [1.4.0]: https://github.com/mayfield/syndicate/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/mayfield/syndicate/compare/v1.2.0...v1.3.0