+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 
 ## [Unreleased] - unreleased
+
+
+## [2.1] - 2015-12-04
 ### Added
 - `Service.close` method for freeing resources.
 
@@ -47,7 +50,8 @@
 - First stable release
 
 
-[unreleased]: https://github.com/mayfield/syndicate/compare/v2...HEAD
+[unreleased]: https://github.com/mayfield/syndicate/compare/v2.1...HEAD
+[2.1]: https://github.com/mayfield/syndicate/compare/v2...v2.1
 [2]: https://github.com/mayfield/syndicate/compare/v1.4.0...v2
 [1.4.0]: https://github.com/mayfield/syndicate/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/mayfield/syndicate/compare/v1.2.0...v1.3.0